feat: add info banner to VerticalNav Workspace component and split out that component to its own file / scss files
