--- src/components/VerticalNav/VerticalNav.sass
@@ -1,168 +1,0 @@
-@import '../../styles/_partials/index'
-
-.VerticalNav
-	display: flex
-	flex-shrink: 0
-	flex-direction: column
-	width: 68px
-	height: 100%
-	padding: 30px 0 0
-	@include theme-background-green-else-graydark
-
-	@at-root :global(.__OsWindows .VerticalNav)
-		padding-top: 0
-
-	@at-root :global(.__OsWindows) .VerticalNav
-		padding-top: 0
-
-		@at-root :global(.__OsWindows .VerticalNav .VerticalNavItem_DragRegion),
-		.VerticalNavItem_DragRegion
-			-webkit-app-region: default
-
-	@at-root :global(.VerticalNav .VerticalNavItem_DragRegion),
-	.VerticalNavItem_DragRegion
-		flex-grow: 1
-		-webkit-app-region: drag
-
-	.VerticalNav_Add
-		width: 48px
-		height: 48px !important
-		margin: auto 0 10px 10px
-		padding: 11.5px 0 !important
-		@include if-theme-light()
-			background: $white
-		@include if-theme-dark()
-			background: $green
-		border-radius: 100%
-		transition: background-color .1s ease 0s
-
-		svg
-			width: 25px
-			height: 25px
-			transition: fill .1s ease 0s
-
-			path
-				@include theme-fill-green-else-white
-
-		&:hover
-			background: $green-dark
-
-			svg path
-				fill: $white
-
-	.FlyTooltip.FlyTooltip__AddSite
-		margin-top: -23px
-
-	.VerticalNav_NonNavLinkItemOuter
-		display: flex
-		align-items: center
-		justify-content: center
-
-	.VerticalNav_NonNavLinkItem,
-	a
-		display: flex
-		align-items: center
-		justify-content: center
-		height: 62px
-		padding: 15px 0
-		text-align: center
-
-		svg path
-			transition: fill .1s ease 0s
-
-		&:global(.__Active)
-			@include if-theme-light()
-				background: $green-dark
-			@include if-theme-dark()
-				background: none
-
-			svg path
-				fill: $white !important
-
-		&:hover:not(.__Active) svg path
-			@include theme-fill-greendark50-else-white-hover
-
-	svg
-		width: 32px
-
-		path
-			@include if-theme-light()
-				fill: $green-dark
-			@include if-theme-dark()
-				fill: $gray75
-
-	.VerticalNavItem
-		text-align: center
-
-		@at-root :global(.VerticalNavItem_Addons)
-			svg
-				width: 31px
-
-.WorkspaceSwitcher
-	.WorkspaceSwitcher_Popup__Width100
-		width: 100%
-
-	.WorkspaceSwitcher_PopupGrid
-		display: grid
-		grid-template-columns: 1fr 1fr 1fr // 3 columns of equal width using 'fr' to ensure 100% when using gaps
-		grid-row-gap: 20px
-		row-gap: 20px
-		grid-column-gap: 20px
-		column-gap: 20px
-		padding: 20px 20px 15px 20px
-
-	.WorkspaceSwitcher_PopupGridItem
-		cursor: pointer
-
-		&::after
-			content: ''
-			position: absolute
-			top: 0
-			right: 0
-			bottom: 0
-			left: 0
-			margin: -2px
-			border-radius: 8px
-			border: 2px solid $gray15
-			@include if-theme-dark
-				border-color: $gray
-
-		&.WorkspaceSwitcher_PopupGridItem__Team::after
-			border-radius: 50%
-
-	.WorkspaceSwitcher_PopupGridItem.WorkspaceSwitcher_PopupGridItem__Active
-		&::after
-			margin: -2px
-			border: 4px solid $green
-			border-radius: 8px
-
-		&.WorkspaceSwitcher_PopupGridItem__Team::after
-			border-radius: 50%
-
-	.WorkspaceSwitcher_PopupGridItemAdd
-		display: flex
-		align-items: center
-		justify-content: center
-		border-radius: 50%
-		border: 2px solid $gray75
-		width: 32px
-		height: 32px
-
-		svg
-			width: 12px
-			stroke: $gray75
-			stroke-width: 2px
-
-	.WorkspaceSwitcher_PopupFooter
-		white-space: nowrap
-		padding: 15px 20px 15px 20px
-
-	.WorkspaceSwitcher_PopupFooter > button
-		display: block
-		padding: 0
-
-		& + button
-			margin-top: 10px
-
-	.WorkspaceSwitcher_Avatar img
-		background: $white
