+++ src/components/VerticalNav/components/WorkspaceSwitcher.scss
@@ -1,0 +1,124 @@
+@import '../../../styles/_partials/index';
+
+.WorkspaceSwitcher {
+	.WorkspaceSwitcher_Popup__Width100 {
+		width: 100%;
+	}
+
+	.WorkspaceSwitcher_Popup_InfoContainer {
+		display: flex;
+		align-items: center;
+		margin: 10px;
+		padding: 10px;
+		color: $white;
+		font-weight: $font-weight-700-bold;
+		background: $green;
+		border-radius: 4px;
+	}
+
+	.WorkspaceSwitcher_Popup_WarningSvg {
+		width: 18px;
+		height: 18px;
+		margin-top: -2px;
+		vertical-align: middle;
+
+		path, circle, rect {
+			fill: $white;
+		}
+	}
+
+	.WorkspaceSwitcher_Popup_CloseSvg {
+		width: 8px;
+		height: 8px;
+		margin-top: -2px;
+		vertical-align: middle;
+
+		path, circle, rect {
+			fill: $white;
+		}
+	}
+
+	.WorkspaceSwitcher_Popup_Text {
+		flex-grow: 1;
+		text-decoration: underline;
+	}
+
+	.WorkspaceSwitcher_PopupGrid {
+		display: grid;
+		grid-template-columns: 1fr 1fr 1fr; // 3 columns of equal width using 'fr' to ensure 100% when using gaps
+		grid-row-gap: 20px;
+		row-gap: 20px;
+		grid-column-gap: 20px;
+		column-gap: 20px;
+		padding: 0 20px 15px 20px;
+	}
+
+	.WorkspaceSwitcher_PopupGridItem {
+		cursor: pointer;
+
+		&::after {
+			content: '';
+			position: absolute;
+			top: 0;
+			right: 0;
+			bottom: 0;
+			left: 0;
+			margin: -2px;
+			border-radius: 8px;
+			border: 2px solid $gray15;
+			@include if-theme-dark {
+				border-color: $gray;
+			}
+		}
+
+		&.WorkspaceSwitcher_PopupGridItem__Team::after {
+			border-radius: 50%;
+		}
+	}
+
+	.WorkspaceSwitcher_PopupGridItem.WorkspaceSwitcher_PopupGridItem__Active {
+		&::after {
+			margin: -2px;
+			border: 4px solid $green;
+			border-radius: 8px;
+		}
+
+		&.WorkspaceSwitcher_PopupGridItem__Team::after {
+			border-radius: 50%;
+		}
+	}
+
+	.WorkspaceSwitcher_PopupGridItemAdd {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		width: 32px;
+		height: 32px;
+		border-radius: 50%;
+		border: 2px solid $gray75;
+
+		svg {
+			width: 12px;
+			stroke: $gray75;
+			stroke-width: 2px;
+		}
+	}
+
+	.WorkspaceSwitcher_PopupFooter {
+		white-space: nowrap;
+		padding: 15px 20px 15px 20px;
+	}
+
+	.WorkspaceSwitcher_PopupFooter > button {
+		display: block;
+		padding: 0;
+
+		& + button {
+			margin-top: 10px;
+		}
+	}
+
+	.WorkspaceSwitcher_Avatar img {
+		background: $white;
+	}
+}
