+++ src/index.ts
@@ -51,7 +51,8 @@ export {default as TableListRepeater} from './components/TableListRepeater';
 export {default as TabNav} from './components/TabNav';
 export {TertiaryNav as TertiaryNav, TertiaryNavItem as TertiaryNavItem} from './components/TertiaryNav';
 export {default as Truncate} from './components/Truncate';
-export {VerticalNav as VerticalNav, VerticalNavItem as VerticalNavItem, WorkspaceSwitcher as WorkspaceSwitcher} from './components/VerticalNav';
+export {VerticalNav as VerticalNav, VerticalNavItem as VerticalNavItem} from './components/VerticalNav';
+export {WorkspaceSwitcher} from './components/VerticalNav/components/WorkspaceSwitcher';
 export {VirtualList} from './components/VirtualList';
 export {VirtualTable} from './components/VirtualTable';
 export {default as WindowsToolbar} from './components/WindowsToolbar';
